# @FIR-1006 - GGML: PERF changes with following option