./tsi-pkg-build.sh "debug"
drwxr-xr-x  3 akapoor tsiusers     4096 Oct  6 22:02 tsi-ggml
-rw-r--r--  1 akapoor tsiusers 16596324 Oct  6 22:02 tsi-ggml-0.0.9.tz
-rw-r--r--  1 akapoor tsiusers        0 Oct  6 22:08 tsi-op.txt
-rw-r--r--  1 akapoor tsiusers   216208 Oct  6 22:08 ggml_perf-all-shape.log
[akapoor@wssw01 llama.cpp]$ 

POsix  result
[akapoor@wssw01 llama.cpp]$   build-posix/bin/llama-cli -p "my cat's name" -m /proj/rel/sw/ggml/models/Tiny-Llama-v0.3-FP32-1.1B-F32.gguf --device tSavorite -c 12288 --temp 0.0 --n-predict 10 --repeat-penalty 1.5 -b 1024 --top-k 50 --top-p 0.9 --repeat-last-n 5 --no-warmup --no-display-prompt
 is Luna.
I'm a cat



llama_perf_sampler_print:    sampling time =      24.92 ms /    16 runs   (    1.56 ms per token,   642.11 tokens per second)
llama_perf_context_print:        load time =    2926.90 ms
llama_perf_context_print: prompt eval time =    2318.16 ms /     6 tokens (  386.36 ms per token,     2.59 tokens per second)
llama_perf_context_print:        eval time =    3709.19 ms /     9 runs   (  412.13 ms per token,     2.43 tokens per second)
llama_perf_context_print:       total time =    6663.61 ms /    15 tokens

=== GGML Perf Summary ===
  Op               Target      Runs        Total us            Avg us
  ADD              OPU         2024         3127498           1545.21
  MUL              OPU         2070         1265143            611.18
  RMS_NORM         OPU         2070         1277210            617.01
  MUL_MAT          CPU        36433        50387125           1383.01
  CONT             CPU         7572          391048             51.64
  RESHAPE          CPU        11215            5531              0.49
  VIEW             CPU        17361            2274              0.13
  PERMUTE          CPU        13563            1998              0.15
  TRANSPOSE        CPU         3119             732              0.23
  GET_ROWS         CPU          389            1047              2.69
  SET_ROWS         CPU         7143            4922              0.69
  SOFT_MAX         CPU         3606          313995             87.08
  ROPE             CPU         7824           33861              4.33
  GLU              OPU         1012         1198366           1184.16

OPU Profiling Results:
------------------------------------------------------------------------------------------------------------------------
Calls  Total(ms)    T/call    Self(ms)  Function




Will testing for below and add the result soon
############
./tsi-pkg-build.sh 
POSIX
FPGA
##########
./tsi-pkg-build.sh "release"
POSIX
FPGA